requirements.txt
nextcord==2.6.0
Unidecode==1.3.8
openai==1.59.6
requests==2.32.3
setuptools==75.6.0
pillow==11.0.0
python-dotenv==1.0.1
ploty==5.24.1
kaleido==0.2.1